Add -[FBProcessFetcher launchdProcessesToContainingDeviceSet]
Summary: This will return a mapping of `launchd_sim` Process to the Device Set that contains the Simulator corresponding to the `launchd_sim`. This can be used to find the Device Sets that correspond to all booted Simulators. This can be useful for cleaning up after any processes that may have booted Simulators in custom sets.

Reviewed By: perryjrandall

Differential Revision: D3722282

fbshipit-source-id: 029e4565151eac7e055f59d032213992f57bf461