Use launchctl to determine Test Runner Process ID
Summary: The method by which we determine the Process Identifier of the Test Runner Host Application doesn't work in all cases as the Binary name and the Process name to not always match. Instead of doing this with Process Querying, we can instead ask the Simulator directly for the Service Name / Process Identifier pair and get a much more consistent result.

Reviewed By: migchez

Differential Revision: D3723028

fbshipit-source-id: c5299c526a6daabe200b55c9aca6a1304d52117f