Add Logic Test integration test.
Summary: A Logic Test can run without a booted Simulator. In order to do this, we need a path to an `otest-shim` which will provide structured output to `stdout` that is then read by `fbxctest` and provided to the internal reporting system. With this in mind, we can run an integration test against the Fixture test bundles and confirm the reporting is correct.

Reviewed By: k21

Differential Revision: D3717052

fbshipit-source-id: 9fe4f0741cbb9b2e0e46fabe9429eac7cf002d5e