Add -[FBSimulatorLaunchCtl listServicesWithError:]
Summary: In an effort to get rid of some explicit process fetching, I'm migrating some usages to use `launchctl` directly. Since we already fetch the service list, we can use this to construct a mapping of Service Name to process identifier. This changes the semantics slightly as this does mean we can return services without a process identifier.

Reviewed By: marekcirkos

Differential Revision: D3735146

fbshipit-source-id: 5dca02ac53ff4be4f173d3529e3b8efa9fc77208