Add Code Signature Checking API
Summary:
In Xcode 8 Beta 4 and greater, a code signature is required for `xctest` bundles that are injected into any Application. If this is not confirmed ahead of time, `XCTestBootstrap` will get stuck when attempting to connect to a bundle. Since it is possible to know that this failure *will* happen ahead of time, we can fail-fast in the case that we don't have a code-signature when one is necessary.

Fortunately, we can use the ad-hoc signing identity `-` for doing the signing. This is somewhat documented in the manpage for `codesign(1)`:

> If identity is the single letter "-" (dash), ad-hoc signing is performed.  Ad-hoc signing does not use an identity at all, and identifies exactly one instance of code. Significant restrictions apply to the use of ad-hoc signed code; consult documentation before using This

Reviewed By: k21

Differential Revision: D3741439

fbshipit-source-id: 1dd8c94bdf4572a1bac686ffcbaffaca783877c1