Add code signature checking on test bundle
Summary: Given that we have the ability to check for the existence of a code signature, we can confirm that running on Xcode 8 means that we have a signed bundle. If this doesn't hold true we can then fail fast. This is preferable to stalling the test run waiting to connect to the test bundle, which is rejected for loading due to a lack of a signature.

Reviewed By: k21

Differential Revision: D3741443

fbshipit-source-id: 7abc886f5023376fdb8c4c49f9998b82227e4a4d