Make FBSimulatorFramebuffer optional on Direct Boot
Summary: Previously, a `FBFramebuffer` has always been created when launching with `FBSimulatorLaunchOptionsEnableDirectLaunch`. However, this doesn't need to be the case. When booting a Simulator 'directly' via `CoreSimulator`, a Framebuffer connection is not required. In the case of running Application Tests, a Framebuffer is unecessary unless video needs to be recorded. This means allocating less resources on a host, which is preferable in the Application Tests case.

Reviewed By: marekcirkos

Differential Revision: D3741429

fbshipit-source-id: 0c5ee6dbfec92f1cf4215e2f8a29cc02ecc809c1