Optional Framebuffer for Xcode 8
Summary: I haven't (yet) found a way to make launching via `CoreSimulator`, but without a Framebuffer work for Xcode 7. The same is not true for Xcode 8, where this is absolutely possible. This means that we can make the Framebuffer optional for Xcode 8 by delegating out to the `_Direct_Xcode8` class in the cluster.

Reviewed By: marekcirkos

Differential Revision: D3757252

fbshipit-source-id: 3ad07eeace7e0bc35a2036f906ad9b3daba96a15