Add Crash Reporting to FBDeviceOperator
Summary: The reporting in the case that an Application Test Process crashes isn't great. Fortunately, we have most of the primitives in order to fetch crash log diagnostics for Simulators. By calling out to this API via the `FBDeviceOperator`, we can get more information about the crashing process. Following on from this, I plan to add changes that will allow a 'Result' to be produced and then inspected.

Reviewed By: marekcirkos

Differential Revision: D3763872

fbshipit-source-id: 4fc5aca54904d664f5d581daf929746de80b1c9a