Add FBTestBundleResult
Summary:
Adds a concrete `Result` from an `FBTestBundleResult`. This allows for the 'crashing' case to be concretely encoded.

In Xcode 8 we don't recieve a disconnection notice from the `DTXConnection` of the Bundle, we can't react as easily to a crashed test host. Since I don't currently have a good way of detecting this inside `XCTestBootstrap`, we can delegate this out to the `FBDeviceOperator` periodically on Xcode 8. This polling is unfortunate, but will reliably determine that the host process has crashed. In this eventuallity, we can break out of the waitloop for the test bundle to complete.

Reviewed By: marekcirkos

Differential Revision: D3764512

fbshipit-source-id: 8d1d3d68b9d34fcf218979501955ec70669a12e3