Add FBServiceManagement
Summary:
In order to determine information about the `launchd_sim` process running on the Host, the `FBProcessFetcher` is used. This is preferrable to using a `ps(1)` subprocess, but is still somewhat overkill when the `ServiceManagement` framework can provide the same information. The `launchctl` of the host has all of this information.

It's important to note that many functions in this Framework are marked as 'Deprecated' and to be replaced with `libxpc`. There is currently no replacement at this time. I believe that using `ServiceManagement.framework` is substantially less risky than the very-much undocumented syscalls that are used by `FBProcessFetcher`. `ServiceManagement.framework` is also used internally by `CoreSimulator` for interacting and querying for the `CoreSimulatorService`, so these functions can't go anywhere any-time soon. Since Xcode 8's `CoreSimulator` works on 10.12, it stands to reason that `ServiceManagement.framework` is currently functional (but still deprecated). When a replacement in `libxpc` is given, we can start using that instead.

Reviewed By: marekcirkos

Differential Revision: D3763453

fbshipit-source-id: eda8157bc4763e3ca7cf61013792085641b96381