Add FBSimulatorTestRunStrategy
Summary: Extracts out the Simulator-to-XCTestBootstrap bridging code that has previously resided in `FBSimulatorInteraction+XCTest`. This means that this class can be used directly by `fbxctest` in order to get the error itself out. The `FBSimulatorInteraction` is sub-optimal in this situation as the value returned from the test run can be significant and interesting.

Reviewed By: marekcirkos

Differential Revision: D3772472

fbshipit-source-id: 9251533b7329bff6f2c6e2f4aab54207cf523c40