Improve teardown semantics for Bundle & Daemon
Summary:
With the legacy protocol, there might not be a guarantee that the daemon disconnection notification will be sent. Therefore we need to manually disconnect if we reach the end of the test plan. In order to handle this, we need to do some case analysis on the state of the daemon/bundle and adjust the result accordingly.

There were also a few bits of descriptions that could use some adjustment to improve the clarity of the logging. `-[FBTestBundleResult disconnect]` was also modified to return a value like in `FBTestDaemonResult`.

Reviewed By: nqmtuan

Differential Revision: D3791071

fbshipit-source-id: 5d693a9fc2958ed3b9fb8a0d3db1e85777a9a07b