Fix 'Unclaimed' Application Fetching
Summary:
There is one very subtle case that will cause the 'unclaimed' (read Xcode-launched) Application fetching will fail:
1) Boot a Simulator via conventional means like via Xcode or an alias to the `Simulator.app`
2) Boot another Simulator.app via `fbsimctl`
3) `fbsimctl shutdown` will properly kill the sim from #1 but not #2

This is because the booting of #2 overwrites the `NSUserDefaults` from #1 so the 'unclaimed' simulator container app doesn't correlate. The check for the UDID in the `NSUserDefaults` is unecessary since we can reasonably assume if there is one more Simulator app that is unclaimed this must be the sole remaining Simulator.app for the `launchd_sim`.

Reviewed By: marekcirkos

Differential Revision: D3811248

fbshipit-source-id: e9f0fb97c8370f5966f191eebbd246f61ec722b5