Don't abort if a OS Version/Device Configuration is missing
Summary: This causes a lot of update pain when Apple changes these in different versions of Xcode. Instead of aborting, we should really just log this fact instead when building the default set, so that the changes can be logged later. I will also be able to add a test in future to confirm that the absent device/os versions are not missing.

Reviewed By: marekcirkos

Differential Revision: D3835356

fbshipit-source-id: 39c90f554f6d377684434ae9aab10f2ed4226637