Allow post-boot connection of Framebuffer in Xcode 8
Summary: In Xcode 8, it becomes possible to obtain the `IOSurface` used by a Simulator and connect to it. This means that the creation of the `FBFramebuffer` becomes deferrable and also opens up the opportunity to start a video encoder from a different process to the one that booted the Simulator.

Reviewed By: marekcirkos

Differential Revision: D3855676

fbshipit-source-id: c7791a5547cad83a57c880cd49aaeb08d05e1fd8