Make EventReporterSubject a first-class protocol
Summary:
By promoting the `EventReporterSubject` to a first-class type from a `typealias` of protocol composition, it becomes possible to have polymorphic protocol extensions as well as default implementations of methods with protocol extensions.

This will allow `Subjects` to be joined and otherwise combined to produce better output for the CLI and HTTP Server. These manipulations will make it substantially easier to improve event reporting with simple combinators

Reviewed By: marekcirkos

Differential Revision: D3862472

fbshipit-source-id: b9cd59a1286f02005051515202015945acea00cd