Add ability to return a result from CommandResult
Summary:
I've played with improving the notions of what should and shouldn't be reported to humans by default vs. the firehose that is the json reporter. I think the best alternative here is to work with the `CommandResult` to support the returning of a value from running some interaction, this then means that it's up to the caller of the `Runner` to determine how to report the result.

The result is that now we have a way of representing all of this inside a `Subject`. As `Subjects` are a first-class protocol, it becomes easier to define how results can be combined. The first example of this is the `list` command, which now returns a `Subject` rather than reporting it. This now means that the caller will perform the reporting at the root. This now means that the result of an `Action` can be put inside the Http Request, rather than injecting a reporter that sucks up all the events.

The new `CompositeSubject` becomes a way of joining subjects together so that they can be all reported at once, in the same way as would have been done with calling `report` for each of the sub-subjects.

Reviewed By: marekcirkos

Differential Revision: D3869679

fbshipit-source-id: cf7958d7834f25ad554ebf5c025b6ed93a517c25