Return 'subject' from Http Relay
Summary:
Now that it's possible for a result to return a value in the Success case, this value can be extracted and serialized to JSON. This makes the case for a number of `GET` requests more compelling and means that the events do not need to be inspected to extract out this information.

Currently, `list` is the only command which returns a value inside the `CommandResult` enum. This means that now the `list` endpoint can use this *explicitly* in the `subject` rather than scouring through the big list of events. This should make returning values from HTTP Requests far easier.

Reviewed By: marekcirkos

Differential Revision: D3869687

fbshipit-source-id: 5fe062b3cf7af544fd657702e2d0bd661f608299