Extract List to a separate runner.
Summary: This will allow the Http Relay to always return the list of Simulators inside an Array, even when there is only one element. Additionally it makes the implementation of this command clearer, as it doesn't become necessary to make a runner for each item that is returned in the query.

Reviewed By: nqmtuan

Differential Revision: D3876373

fbshipit-source-id: 710f15e7416d562dd207b88fd03c9c03dc9237a8