Add support for GET requests in HTTP Relay
Summary: Currently, all of the HTTP Requests are `POST` due to a limitation in the old HTTP Stack. Now that this is no longer the case, we can define routes for different HTTP Methods. Additionally, using a regex on the path of the request means that all incoming requests can define a query in the path of the request by passing a UDID in the path. This makes it much easier to target the same device in a client by just concatenating strings in a path.

Reviewed By: marekcirkos

Differential Revision: D3876375

fbshipit-source-id: 2b4fafeafa8d56b55c40fd88369a31cdb95606b9