Add 'config' Action
Summary:
This is a command that can be used for learning useful information about the current environment.

For example, the output of this command can be used to determine whether ad-hoc video recording is supported. This allows callers of the `fbsimctl` command to know which features will and won't work before attempting to use them.

Reviewed By: marekcirkos

Differential Revision: D3876381

fbshipit-source-id: 92b62dd32a0c3fd777ac33bf87cfaa8d3fed5f69