Make Application Bundle ID Optional for clear_keychain
Summary: The `clear_keychain` doesn't need to take into account a running application, as the method for destroying the keychain now relies on a system-wide service. Instead of repeating code in the body of the interaction used by `clear_keychain` we can instead use an existing interaction to do this.

Reviewed By: migchez

Differential Revision: D3892595

fbshipit-source-id: 0e1a89d77848d24d5a02c0c3229053fdf90e666b