add FBKeychainCleanStrategy
Summary: The current method for removing the data from a keychain is to inject a shim into a running application and use this to empty the contents of the keychain. This is probably very overkill for what we want, since we just want to be able to remove all the contents of a keychain to support a way of getting the Simulator into a known state without the need to restart.

Reviewed By: migchez

Differential Revision: D3892584

fbshipit-source-id: a8a63b5e486344196056839769c46ad3215c105d