Migrate to Swift 3
Summary: There's now no reason to hold off on using Xcode 8 as this is the current latest version of Xcode. `FBSimulatorControl` the Framework will still be buildable on Xcode7, but there are breaking changes in Swift that means we can only build `fbsimctl` with Xcode 8. I'll update the travis build matrix to invert the `fbsimctl` inclusion & exclusion logic.

Reviewed By: marekcirkos

Differential Revision: D3899604

fbshipit-source-id: e92114d972526cf53e5c0ca24c8894ab8a39e0af