Fix missing -[SimRuntime supportedRuntimes] & -[SimDeviceType supportedDeviceTypes]
Summary:
In Xcode 8.1 Beta, these methods have gone from `SimRuntime` and `SimDeviceType` since the correct way is to use the `SimServiceContext`. Fortunately, I recently made a `FBSimulatorServiceContext` class that can emulate the old way of fetching these arrays as well as a new way that uses `SimServiceContext`.

I've had to change the initialization of `FBSimulatorServiceContext` to be a Singleton so it is globally accessible, but this shouldn't pose a problem as it wraps the singleton `SimServiceContext` regardless.

Reviewed By: marekcirkos

Differential Revision: D3906614

fbshipit-source-id: f327f10510fbf91de2bdcd104d93c6b8225bb595