Promote FBInteraction
Summary:
I think I've finally figured out a way of extracting the desirable behaviour of the interaction model (convenience) whilst decomposing the implementation of the interacton itself into concrete classes. The `FBInteraction` API provides a reciever that has closed over all of it's necessary arguments in order to perform some task. This means that the majority of all the strategies can be adapted to become interactions, rather than relying on glue categories to wire them up.

This also means that the implementation of the command line runner can provide a mapping of the `Action` enum to implementations of Interactions. This will also accomodate implementations for `FBDevice` which can provide similar `FBInteraction` implementations without requiring the chainable API that has proven to be of little worth.

Reviewed By: migchez

Differential Revision: D3892601

fbshipit-source-id: 3220c5a4565262fbb6f4945b1e5ff0c3e246a69c