Add launchApplication to FBApplicationCommands
Summary:
This will consolidate the interface for launching an Application across devices and Simulators, which will also accomodate `fbsimctl` interacting with `FBSimulator` and `FBDevice` in the same way.

There are some intentional changes to the way that information about an app launch is passed around inside `XCTestBootstrap` this is intentional since the model object is constructed eventually anyway, the arguments are typed appropriately and the wrapper object provides descriptions.

Reviewed By: marekcirkos

Differential Revision: D3913280

fbshipit-source-id: dd7eec5cc2c4e4b5db293ec6c77c96ec0349ea0f