FBTestRunConfiguration -> FBXCTestConfiguration
Summary: This will help disambiguate the roles of configuration objects between `fbxctest` and `FBXCTestBootstrap`. In future the `FBXCTestConfiguration` will be broken down into smaller parts for each of the kinds of test runs. Additionally, I'd like to be able to serialize the configuration so that it can be represented in a number of different forms beyond being parsed from command line arguments.

Reviewed By: marekcirkos

Differential Revision: D3913286

fbshipit-source-id: c16ef47b5787692982906dbc4e519b824f286949