Adding framework to capture CLI argument parser formats
Summary:
This diff introduces a framework that can be used to annotate the `fbsimctl` CLI argument parser with metadata (`ParserDescription`s) that can then be used to generate a help/usage dialog.

I have modified the current help dialog to show an example of how the metadata is formatted into a Help dialog (I will update the example with something more substantial shortly), before this diff is landed, this will be removed.

The metadata is represented as a tree, roughly corresponding to a regular grammar in which certain nodes are annotated with usage information. Special section nodes can be used to break up longer formats into logically related pieces. When formatting the help dialog, we only print information about each section and each primitive once, and we only print information about each flag once per section.

The purpose of this diff is to nail down details about how the dialog should be formatted. For example, currently I use square brackets to wrap around section tags that appear in other descriptions and to wrap around formats that need to be delimited (For example `[--foo | --bar]`) as they appear nested within other formats. Perhaps we could use a different syntax for one or the other?

Differential Revision: D3950909

fbshipit-source-id: 93f8e8316a1fa76f2eba4ccd6a87baa1fed24502