Replacing String-based parser descriptions with `ParserDescription`.
Summary:
Looking at the output (in the test plan) this diff doesn't seem to improve the situation very much. The plan is, in another stacked diff, I will split the parser descriptions into logical sections, which will make the usage string much more intelligible. The motivation for this diff is just to connect everything up, without making anything worse.

In this diff we change the type of the `matchDescription` parameter in the `Parser` struct from `String` to `ParserDescription` which we created in the diff this depends on.

Knock-on effects of this change:

 -  Parser combinators need to combine descriptions as well as parsing functions.
 -  Parsers generally must be augmented with annotations .

I have completed the former in this diff, but when it comes to the latter, I have just written generic explanations of commands. These will need to be fleshed out fully in a subsequent diff.

Differential Revision: D3960974

fbshipit-source-id: f70fe68c68045753e8296bfd105875a18305ece0