Splitting Argument format into sections.
Summary:
Annotate the descriptions of parsers with sections to make formats more intelligible.

Again, the explanations I've put in here don't mean anything. The parts I can actually explain, I will in a diff to come. In order for me to fill out the others myself, I would have to get in touch with someone from `fbsimctl` who can explain them to me. This is probably not the best use of everybody's time, however, so I'm hoping that the description framework is simple enough that I can hand-over the task of filling that stuff in to someone with more domain specific knowledge.

Differential Revision: D3967637

fbshipit-source-id: 082c000b3a7c89d99034e9deb34ac58141af2900