Add fbsimrecord wrapper script
Summary:
In order to accomodate some common workflows with conveniences, I've added a `Scripts` directory for the cli apps. This means that a script can be committed to the relevant directory and the default installation of the clis will include the scripts with them. As a result, installing via homebrew will result in these scripts being on the `$PATH` so that they can be easily recalled and shell-completed.

The first of these scripts is `fbsimcrecord`, which is a simple way of using `fbsimctl` to record the videos of all booted simulators. Very convenient for demoing out features and showing repros for bugs.

Reviewed By: asm89

Differential Revision: D4031531

fbshipit-source-id: 3f4722bf725c3d828fbd69d5abed33a3ca72e085