Enforce requirement that Dictionary/Array is at the root level of serialization
Summary: `JSON` Serialization can break if the root node passed in is not a Dictionary or Array. For this reason, we should check that there is a container at the root level before passing to `NSJSONSerialization` otherwise it is possible to hit an assertion inside of `Foundation`. This will ensure that we do not hit this trap in the first place by only allowing `NSArray`/`NSDictionary` to be serialized. Strings, Numbers etc. cannot be.

Reviewed By: asm89

Differential Revision: D4036423

fbshipit-source-id: ae6ec5f41d4fab260ffb525f6585a2c0ecf4a7aa