Optionally Confirm Shims are Signed
Summary: For each time `fbxctest` is called, it will confirm that the shims are signed using `codesign`. This is massively overkill if you can safely assume that the shims are signed. This is potentially a resource sink when running many tests in parallel. Setting the default to ignore this, but making this opt-in with an environment variable.

Reviewed By: k21

Differential Revision: D4057848

fbshipit-source-id: 4d11d14ae260b95c114ba203c7e06b7fb2df2cf0