Add more required service names for iOS 10
Summary:
Since Xcode 8, and iOS 10, the time between when a device is reported as 'Booted' in `CoreSimulator` and is actually in a usable state has increased on the 'Fresh Boot' path. It appears that a lot of IO happens as there are various bits copied from Xcode into the Simulator directory.

The problem this poses, is it can mask the timings for certain activities. For, example if calling `launch_xctest` immediately after 'boot' returns, there will be a long wait for the Simulator to actually be in a state where the app launch will actually happen. This means that the `launch_xctest` command is assuming some of the timing of the 'Booted' Simuator to be usable. In order to make a booted Simulator to be actually usable, I'm adding some more services that should be present before a Simulator is usable.

Reviewed By: marekcirkos

Differential Revision: D4136785

fbshipit-source-id: 7c0a8425658c4b8ffd88f18f23c63185d9ba0038