Replace FBMultiFileReader with FBFileReader
Summary:
The `FBMultiFileReader` exists as a way of multiplexing the reading of pipes and file handles. By using `FBTask` for the subprocess output and a single reader of a file for the output of the shim output file, we can simplify this process. The managment of the File Handle can also then move into the file reader so that the teardown of the open file is tied to the teardown of reading the file. This will allow further consolidation and simplification of file reading in `FBTask` and associated classes.

In addition the `FBLogicTestRunner` now gets `FBTask` semantics. This means that the test can be tested for timing out. Previously `-[NSTask waitUntilExit]` could block the calling thread if the subprocess never terminates.

Reviewed By: marekcirkos

Differential Revision: D4073835

fbshipit-source-id: e66e8fe6a78d2ed257e1e2734e292c7aa9594670