Add sampling of stalled Mac Logic Test
Summary:
Since we launch the `xctest` process for a Mac Logic Test directly, we have the PID of this process after we launch it. If a test process stalls/hangs, we can detect this and then perform a `sample` on it. This `sample` can then be shown in the error message returned to consumers of the test results.

This can only currently be implemented for Mac Logic tests, since we go via `simctl` to launch an `xctest` process. This diff adds the logic for sampling in `FBLogicTestRunner` and a corresponding test.

Reviewed By: marekcirkos

Differential Revision: D4073846

fbshipit-source-id: 304c9e0d6b8981f35c1c163bc4e41bbeb9541409