Add Sampling of stalled iOS Logic Test
Summary: The `sample` of a stalled `xctest` process for a stalled iOS test requires that the `xctest` process is sampled, rather than `simctl`. With this in mind, we can use the `FBProcessFetcher` object to obtain the pid of this process and then sample it as we would a Mac `xctest` process.

Reviewed By: marekcirkos

Differential Revision: D4088083

fbshipit-source-id: 85fb80ec52f0dbc333d01b76e81f04a6d8db3756