One-Stage Configuration Initialization
Summary: Places the loading of a `FBXCTestConfiguration` into a single method, rather than an initializer and a `load` function. Makes it simpler for callers and the error is also propogated appropriately. This will also make it easier in future changes to move relevant configurations into subclasses so that the base confguration doesn't end up having irrelevant properties for the test that is being run. Further decomposing in this way will also make it easier for the configuration to be directly serializable.

Reviewed By: marekcirkos

Differential Revision: D4089298

fbshipit-source-id: d5cf3af28ae1298ae3a387b22fa6f3fee4bddf85