Simplify Simulator Configuration in FBXCTestConfiguration
Summary:
The `FBXCTestConfiguration` has some duplication of Configuration with Simulator Names, as well as the `FBSimulatorConfiguration`. We can instead consolidate everything into a `nullable` `FBSimulatorConfiguration`.

The creation of this configuration is also deferred internally. This means that we only need to create a configuration if we know the configuration represents an iOS Target. This means that in the case that `fbxctest` runs against a Mac, that we never need to load the Frameworks that `FBSimulatorControl` depends on.

Reviewed By: marekcirkos

Differential Revision: D4095132

fbshipit-source-id: 548f7fddb3719d22819e9decb8442c9a021c5880