Subtract launch path from arguments before parsing
Summary: Currently, `fbxctest` is removing the first argument of test runner inside the parser. Instead we should move this outside, so that the Unit Tests for the parser are now operating on the correct inputs. Prior to this change, the unit tests were not operating on the entire array of arguments, which means the tests were not fully correct.

Reviewed By: marekcirkos

Differential Revision: D4197615

fbshipit-source-id: fe280bf2d34d6ce3c870c03f261894e2bc1ca84e