Expose editPropertyListFileRelativeFromRootPath
Summary:
Edit a property list file of the Simulator by manipulating the content of the dictionary in the block.

The `setupKeyboard` method is changing a few of the keyboard settings, but not all of them. With this new method consumers of the Framework have full control of editing property list files that belong to the Simulator. Without the need of FBSimulatorControl having to know what property list files exist and what options they support.
Closes https://github.com/facebook/FBSimulatorControl/pull/338

Differential Revision: D4212844

Pulled By: lawrencelomax

fbshipit-source-id: 061b2461aa9a8c509cc49dd8e85e4460efd9c3e2