Add FBDeviceInflationStrategy
Summary:
Currently, the onus is on the consumers of `FBDevice` to retain references so that `FBDevice` instances are consistent over time. This can be a problem with chaining commands in `fbsimctl`, since the `FBDevice` instances can be requeried at any time, leading to different references.

The `FBDeviceInflationStrategy` will ensure that references are consistent by comparing the state of `FBAMDevice` instances with the current `FBDevice` instances. This can also be optimized to ensure that excessive work to fetch devices is not needed.

Reviewed By: marekcirkos

Differential Revision: D4298697

fbshipit-source-id: 85093fa6d54f0b4d697b3c12232d3504cc690cc0