add install api to http relay
Summary:
this adds a `/install` endpoint to the HTTP api of `fbsimctl`. It expects an IPA file to be posted as the request body. It then
- writes this file to a temporary directory
- extracts the contents
- calls the `install` action on the first directory that's name ends in `.app` from the archive
- cleans up

And it does all this with O(1) memory, for reasons which I'm not 100% sure about :)

For the cleanup part there is a new `PostActionHook` which can be passed to the `install` action. It is only run if install doesn't throw an exception.

Reviewed By: lawrencelomax

Differential Revision: D4468532

fbshipit-source-id: eceec2999985daa3d4315d5ffb0ec83d9b089acd