Use a unique Device Set for Application Tests
Summary: In a prior change, `fbxctest` was switched to re-using Simulators from the Default Device set when running Application Tests. This was important for Xcode 8.1, since it had a substantial performance regression when launching a newly-created Simulator. Thi regression has been fixed in Xcode 8.2, so we can use a Device-Set per Test Process. This restores the behaviour that `fbxctest` can run in parallel on the same host, since each `fbxctest` process can have it's own device set.

Reviewed By: mzlee

Differential Revision: D4501606

fbshipit-source-id: 45c2857896c294bc63e9e232135692c2fc9fb9fd