Fix SimDisplayVideoWriter in Xcode 8.3 Beta 2
Summary:
This API has changed, it now appears that Callback Blocks are used to signal that the video writing has finished. This is likely related to the `simctl` issues that have been present in previous betas where `simctl` would appear to not write out the entirity of the buffer before terminating.

We should use the callback to semaphore the `stopRecording:` `dispatch_group`. However this change at least avoids runtime crashes. I can follow up with a change to improve the teardown path for videos.

Differential Revision: D4522163

fbshipit-source-id: 3a00bf3101bbee03f8ec8b28a8e9638bfc3c616e