Add sendHomeButton API to FBSimulatorHID
Summary:
Adds a provisional API for sending a Home Button using `Indigo`. This is based on the `Button` payload that we previously have more information for. There is an internal implementation that we can later make public for different kinds of payloads. As it stands, this can be re-used for the touch up and the touch down. A 'tap' is a higher-level API for what is fundementally a touch-down followed by a touch-up.

For more information about the reversing of this protocol, consult the Documentation directory. There is also some documentation of what the fields are expected to be, in the implementation of `FBSimulatorHID`

Reviewed By: asm89

Differential Revision: D4496268

fbshipit-source-id: 33510638f62132e47ddfff8c47aeab3ded3bea4d