Add Tap Events
Summary:
Adds an API for tapping at a point, using the `FBSimulatorHID` class. Just as with the 'Home Button' API, this is formed of a touch-down and then a touch-up. There was a fair bit of monkeying around with the fields and how there is some duplication between the fields. As I work on breaking down the events, this can eventually be exposed as a touch-up and touch-down. For the time being, I'm just exposing the higher-level tap.

There are also some changes to the `backboard_indigo.d`, in order that you can see these events come in at the backend.

Reviewed By: marekcirkos

Differential Revision: D4504259

fbshipit-source-id: dc5e34ba6ac25f9284aad896db19dfbc20f5fbc6