Add FBSimulatorHIDEvent
Summary:
Adds a value-object for containing a HID event. This can be serialized in a manner that allows batching/buffering and remote-transmission.

I've gone down the path of using Composites as much as possible. A 'Short Tap' is formed of a touch-down and then a touch-up immediately after. Long taps can be formed with pauses in-between the events. Building HID Events in this way will mean it will be easy to support generators of events that want to speak at a very immediate level (i.e. they are not making composite events and just translating raw user input), as well as make higher level conveniences by composing a sequence of actions.

Reviewed By: marekcirkos

Differential Revision: D4536298

fbshipit-source-id: 0fab42205bdda9d63b06f23a825067bb4f9ffed5