Add Button to FBSimulatorHIDEvent
Summary:
Expands the Value Representation of a HID interaction, with button presses. The enumeration for mapping to the `Indigo` values is not mirrored in terms of values. This gives us the flexability we need if there are changes to these values in the future, as well as letting us know the true bounds of these enumerated values.

Once again, the previous explicit call in `FBSimulatorHID` has been replaced with the usage of this API for a short press. This means that there should be no changes to the behaviour of this API, it's just moved to a more appropriate place and can be composed of two discrete, sequenced events.

Reviewed By: marekcirkos

Differential Revision: D4536499

fbshipit-source-id: 96129cdea0d048df267ad58cc8635ee35e5ead4c