Migrate /keys to /hid
Summary: We now have an JSON-Serializable Object that can specify HID events. With that in mind, I want to move the API for this at the HTTP-Server level to use `FBSimulatorHIDEvent` as a value. `/keys` is too specific, so we can instead just generate the `/hid` event. If we want `/keys` we can still return an `Action.hid`, except we just extract the relevant values. More general primitives, mean we can build conveniences on top.

Reviewed By: marekcirkos

Differential Revision: D4536598

fbshipit-source-id: 60a13aec4476f6972ce961c9529480f49d054f58