Architecture filtering
Summary:
I followed the path of having an enum for a while, but it turns out to be a pretty deep rabbit hole without any immediate benefit. I decided to go with a C struct, which can be neatly converted into a string-backed enum once the Swift floodgates are open. It does make about 10 lines slightly uglier, and all the properties typed as `NSString *` right now, which is imo worth it.

If you filter via e.g. `--arch=i386`, and don't specify the output format, none of the output includes a mark about the architecture you asked for, because architecture isn't part of the default output format.

Reviewed By: lawrencelomax

Differential Revision: D4544765

fbshipit-source-id: cd4944a258195413ab8086bf1b81b94442ad6037