Expose stat_loc in Agent Launch Callback
Summary:
`simctl` propogates the exit code of the process that is launched when `spawn` is used. We don't currently expose this API, though it is clearly exceedingly useful. For example, we care about the exit code of an `xctest` process if it is spawned using `SimDevice`

This change ensures that this value is accurately typed and that the meaning of the `stat_loc` value is accurately defined

Reviewed By: stepanhruda

Differential Revision: D4579511

fbshipit-source-id: a1b981751d0a68707b530efe3232e2a5923e5f60