Use FBAgentLaunchStrategy for launching xctest
Summary:
Now that we can spawn an agent with file handles for output directly, `fbxctest` can use this to run `-[SimDevice spawnWithPath:options:terminationHandler:error:]`. This removes one-level of complexity in the form of `simctl`, since we can directly spawn the `xctest` executable without going through `simctl` first.

This will also enable easier debugging of the `xctest` process, as the spawned `xctest` process can be launched with `wait_for_debugger`. When spawning the `xctest` process, we will get the `pid` which can be back-propogated the the concerned listener, for forwarding onto the debugger.

Reviewed By: marekcirkos

Differential Revision: D4572078

fbshipit-source-id: 7b0d1fbf158021de8a952a9668d55e190279078a