Mirror existing xctool -waitForDebugger flag
Summary: Turns out xctool already has a flag like this! I didn't mirror the output JSON event, in xctool it's under begin-status, imo it's much better to have a custom event type we can direclty extract the PID from. For xctool we'll need to parse the message to get the pid...

Reviewed By: lawrencelomax

Differential Revision: D4591777

fbshipit-source-id: 453bc8e5fe7ac2a180c457bb8ad9466692b88e2d