Use NS_STRING_ENUM in FBiOSTargetFormat
Summary: Now I'm about to remove building on Xcode 7, we can start to use some of the Objective-C/Swift affordances that were added in Xcode7/Swift3. First up is `NS_STRING_ENUM` usage in `FBiOSTargetFormat`. This means that the keys will be typed to live innside a closed-enum, rather than being completely stringly-typed.

Reviewed By: marekcirkos

Differential Revision: D4597464

fbshipit-source-id: 2516fb9e9b233afdae1bff552384202b5b010b56