Pass down FBFramebufferRenderable
Summary: I want to keep the classes having a 'known' public `FBSimulatorControl` class if it is part of the API and hiding the implementation details of `CoreSimulator` and `SimulatorKit`. As `CoreSimulator` isn't exported as a module, consumers won't be able to understand the available APIs for `CoreSimulator`. These should be wrapped and abstracted by `FBSimulatorControl`

Reviewed By: marekcirkos

Differential Revision: D4649558

fbshipit-source-id: 4e86c399a91de5192af12c228b2be3f5a2f3e811