Add a Surface Consumer Protocol
Summary: Adds a Public Protiocol for consuming a Surface. Over time I want to make this a Public interface that can be used outside of `FBSimulatorControl`, so anyone can integrate with the Surface API. This will make it substantially easier for 3rd-parties to provide their own Framebuffer Sinks, outside of whatever exists in `FBSimulatorControl` at the time.

Reviewed By: marekcirkos

Differential Revision: D4658362

fbshipit-source-id: b1d70cb5303523e06330fdc089a672b1e135fba3