Registration-based FBFramebufferFrameGenerator
Summary:
This change will invert the connection between Frame Sinks and the Frame Generator.

This means that it will be possible to attach Sinks at any time and avoid the need for any listening to be set up in advance. The listening can be made lazy, so that interested parties can be notified only when it becomes necessary for a connection to be made. As a result, any launch that connects a Framebuffer but doesn't do anything with the Framebuffer will not result in Frames being created, saving work.

Reviewed By: marekcirkos

Differential Revision: D4659524

fbshipit-source-id: b57ed94ac31c5e4e2dad0eecb074069f3af2f745